Move auth to a util to start refactor
